@@ -20,17 +20,20 @@
     "irc": "irc://irc.freenode.net/imbo"
   },
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.4 | >=8.0",
+    "ext-fileinfo": "*",
+    "ext-json": "*",
     "behat/behat": "^3.0",
-    "guzzlehttp/guzzle": "^6.0",
+    "guzzlehttp/guzzle": "^7.3.0",
     "beberlei/assert": "^2.1",
-    "firebase/php-jwt": "^4.0 | ^5.0"
+    "firebase/php-jwt": "^5.0"
   },
   "require-dev": {
     "silex/silex": "^2.0",
     "symfony/security": "^3.1",
     "symfony/process": "^3.1",
-    "phpunit/phpunit": "^5.5"
+    "phpunit/phpunit": "^9.5",
+    "ext-sockets": "*"
   },
   "autoload": {
     "psr-4": {
