@@ -1,29 +1,34 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Context\Initializer;
 
-use Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
-use Imbo\BehatApiExtension\ArrayContainsComparator as Comparator;
-use Imbo\BehatApiExtension\Context\ArrayContainsComparatorAwareContext;
 use Behat\Behat\Context\Context;
 use Behat\Behat\Context\Initializer\ContextInitializer;
+use Imbo\BehatApiExtension\ArrayContainsComparator as Comparator;
+use Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
+use Imbo\BehatApiExtension\Context\ArrayContainsComparatorAwareContext;
 
 /**
  * Array contains comparator context aware initializer
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class ArrayContainsComparatorAwareInitializer implements ContextInitializer {
+class ArrayContainsComparatorAwareInitializer implements ContextInitializer
+{
     /**
      * @var Comparator
      */
-    private $comparator;
+    private Comparator $comparator;
 
     /**
      * Class constructor
      *
      * @param Comparator $comparator
      */
-    public function __construct(Comparator $comparator) {
+    public function __construct(Comparator $comparator)
+    {
         $comparator
             ->addFunction('arrayLength', new Matcher\ArrayLength())
             ->addFunction('arrayMinLength', new Matcher\ArrayMinLength())
@@ -42,7 +47,8 @@ class ArrayContainsComparatorAwareInitializer implements ContextInitializer {
      *
      * @param Context $context
      */
-    public function initializeContext(Context $context) {
+    public function initializeContext(Context $context): void
+    {
         if ($context instanceof ArrayContainsComparatorAwareContext) {
             $context->setArrayContainsComparator($this->comparator);
         }
