@@ -1,15 +1,20 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use Imbo\BehatApiExtension\ArrayContainsComparator;
 use Imbo\BehatApiExtension\ArrayContainsComparator\Exception\ArrayContainsComparatorException;
-use PHPUnit_Framework_TestCase;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\JWT
  * @testdox JWT matcher
  */
-class JWTTest extends PHPUnit_Framework_TestCase {
+class JWTTest extends TestCase
+{
     /**
      * @var JWT
      */
@@ -18,7 +23,8 @@ class JWTTest extends PHPUnit_Framework_TestCase {
     /**
      * Set up matcher instance
      */
-    public function setup() {
+    public function setup(): void
+    {
         $this->matcher = new JWT(new ArrayContainsComparator());
     }
 
@@ -27,15 +33,16 @@ class JWTTest extends PHPUnit_Framework_TestCase {
      *
      * @return array[]
      */
-    public function getJwt() {
+    public function getJwt()
+    {
         return [
             [
                 'jwt' => 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ',
                 'name' => 'my jwt',
                 'payload' => [
                     'sub' => '1234567890',
-					'name' => 'John Doe',
-					'admin' => true,
+                    'name' => 'John Doe',
+                    'admin' => true,
                 ],
                 'secret' => 'secret',
             ],
@@ -51,22 +58,27 @@ class JWTTest extends PHPUnit_Framework_TestCase {
     }
 
     /**
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage No JWT registered for "some name".
+     *
+     *
      * @covers ::__invoke
      */
-    public function testThrowsExceptionWhenMatchingAgainstJwtThatDoesNotExist() {
+    public function testThrowsExceptionWhenMatchingAgainstJwtThatDoesNotExist()
+    {
+        $this->expectExceptionMessage("No JWT registered for \"some name\".");
+        $this->expectException(InvalidArgumentException::class);
         $matcher = $this->matcher;
         $matcher('some jwt', 'some name');
     }
 
     /**
-     * @expectedException Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException
+     *
      * @expectedExceptionMessage Haystack object is missing the "some" key.
      * @covers ::addToken
      * @covers ::__invoke
      */
-    public function testThrowsExceptionWhenJwtDoesNotMatch() {
+    public function testThrowsExceptionWhenJwtDoesNotMatch()
+    {
+        $this->expectException(Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException::class);
         $matcher = $this->matcher->addToken('some name', ['some' => 'data'], 'secret', 'HS256');
         $matcher(
             'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ',
@@ -83,7 +95,8 @@ class JWTTest extends PHPUnit_Framework_TestCase {
      * @param array $payload
      * @param string $secret
      */
-    public function testCanMatchJwt($jwt, $name, array $payload, $secret) {
+    public function testCanMatchJwt($jwt, $name, array $payload, $secret)
+    {
         $matcher = $this->matcher->addToken($name, $payload, $secret);
         $matcher(
             $jwt,
@@ -97,10 +110,14 @@ class JWTTest extends PHPUnit_Framework_TestCase {
      * @covers ::__construct
      * @covers ::__invoke
      */
-    public function testThrowsExceptionWhenComparatorDoesNotReturnSuccess() {
-        $comparator = $this->createConfiguredMock(ArrayContainsComparator::class, [
-            'compare' => false,
-        ]);
+    public function testThrowsExceptionWhenComparatorDoesNotReturnSuccess()
+    {
+        $comparator = $this->createConfiguredMock(
+            ArrayContainsComparator::class,
+            [
+                'compare' => false,
+            ]
+        );
         $matcher = (new JWT($comparator))->addToken(
             'token',
             [
