@@ -1,22 +1,31 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Context\Initializer;
 
-use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
 use GuzzleHttp\Client;
-use PHPUnit_Framework_TestCase;
+use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
+
+use function socket_create;
 
 /**
- * @coversDefaultClass Imbo\BehatApiExtension\Context\Initializer\ApiClientAwareInitializer
+ * @coversDefaultClass \Imbo\BehatApiExtension\Context\Initializer\ApiClientAwareInitializer
  * @testdox Initializer for API Client aware contexts
  */
-class ApiClientAwareInitializerTest extends PHPUnit_Framework_TestCase {
+class ApiClientAwareInitializerTest extends TestCase
+{
     /**
      * @covers ::initializeContext
      * @covers ::validateConnection
-     * @expectedException RuntimeException
-     * @expectedExceptionMessage Can't connect to base_uri: "http://localhost:123".
      */
-    public function testThrowsExceptionWhenBaseUriIsNotConnectable() {
+    public function testThrowsExceptionWhenBaseUriIsNotConnectable(): void
+    {
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage("Can't connect to base_uri: \"http://localhost:123\".");
+
         $initializer = new ApiClientAwareInitializer(['base_uri' => 'http://localhost:123']);
         $initializer->initializeContext($this->createMock(ApiClientAwareContext::class));
     }
@@ -26,13 +35,15 @@ class ApiClientAwareInitializerTest extends PHPUnit_Framework_TestCase {
      * @covers ::validateConnection
      * @covers ::__construct
      */
-    public function testInjectsClientWhenInitializingContext() {
+    public function testInjectsClientWhenInitializingContext(): bool
+    {
         // Set up a socket for the test case, try all ports between 8000 and 8079. If no ports are
         // available the test case will be marked as skipped. This is to get past the base URI
         // validation
-        set_error_handler(function() { return true; });
+        set_error_handler(static fn() => true);
         $sock = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
 
+        $result = false;
         for ($port = 8000; $port < 8079; $port++) {
             if ($result = socket_bind($sock, 'localhost', $port)) {
                 break;
@@ -55,9 +66,11 @@ class ApiClientAwareInitializerTest extends PHPUnit_Framework_TestCase {
             ->method('setClient')
             ->with($this->isInstanceOf(Client::class));
 
-        $initializer = new ApiClientAwareInitializer([
-            'base_uri' => sprintf('http://localhost:%d', $port),
-        ]);
+        $initializer = new ApiClientAwareInitializer(
+            [
+                'base_uri' => sprintf('http://localhost:%d', $port),
+            ]
+        );
         $initializer->initializeContext($context);
 
         // Close socket used in test case
