@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
@@ -8,35 +11,41 @@ use InvalidArgumentException;
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class ArrayMaxLength {
+class ArrayMaxLength
+{
     /**
      * Match the max length of an array
      *
      * @param array $array An array
      * @param int $maxLength The expected maximum length of $array
-     * @throws InvalidArgumentException
+     *
      * @return void
+     * @throws InvalidArgumentException
      */
-    public function __invoke($array, $maxLength) {
+    public function __invoke(array $array, int $maxLength)
+    {
         // Encode / decode to make sure we have a "list"
         $array = json_decode(json_encode($array));
 
         if (!is_array($array)) {
-            throw new InvalidArgumentException(sprintf(
-                'Only numerically indexed arrays are supported, got "%s".',
-                gettype($array)
-            ));
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Only numerically indexed arrays are supported, got "%s".',
+                    gettype($array)
+                )
+            );
         }
 
-        $maxLength = (int) $maxLength;
         $actualLength = count($array);
 
         if ($actualLength > $maxLength) {
-            throw new InvalidArgumentException(sprintf(
-                'Expected array to have less than or equal to %d entries, actual length: %d.',
-                $maxLength,
-                $actualLength
-            ));
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Expected array to have less than or equal to %d entries, actual length: %d.',
+                    $maxLength,
+                    $actualLength
+                )
+            );
         }
     }
 }
