@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
@@ -8,36 +11,45 @@ use InvalidArgumentException;
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class GreaterThan {
+class GreaterThan
+{
     /**
      * Match a numeric value
      *
      * @param numeric $number A variable
      * @param numeric $min The minimum value of $number
-     * @throws InvalidArgumentException
+     *
      * @return void
+     * @throws InvalidArgumentException
      */
-    public function __invoke($number, $min) {
+    public function __invoke($number, $min)
+    {
         if (!is_numeric($number)) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not numeric.',
-                $number
-            ));
+            throw new InvalidArgumentException(
+                sprintf(
+                    '"%s" is not numeric.',
+                    $number
+                )
+            );
         }
 
         if (!is_numeric($min)) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not numeric.',
-                $min
-            ));
+            throw new InvalidArgumentException(
+                sprintf(
+                    '"%s" is not numeric.',
+                    $min
+                )
+            );
         }
 
         if ($number <= $min) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not greater than "%s".',
-                $number,
-                $min
-            ));
+            throw new InvalidArgumentException(
+                sprintf(
+                    '"%s" is not greater than "%s".',
+                    $number,
+                    $min
+                )
+            );
         }
     }
 }
