@@ -1,7 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Exception;
 
 use Exception;
+use JsonException;
 
 /**
  * Array contains comparator exception
@@ -14,30 +18,36 @@ class ArrayContainsComparatorException extends AssertionFailedException {
      *
      * @param string $message Exception message
      * @param int $code Exception code
-     * @param Exception $previous Previous exception in the stack
-     * @param mixed $needle The needle in the comparison
-     * @param mixed $haystack The haystack in the comparison
+     * @param Exception|null $previous Previous exception in the stack
+     * @param mixed|null $needle The needle in the comparison
+     * @param mixed|null $haystack The haystack in the comparison
+     *
+     * @throws JsonException
      */
-    public function __construct($message, $code = 0, Exception $previous = null, $needle = null, $haystack = null) {
-        // Reusable line of ='s
-        $line = str_repeat('=', 80);
-
+    public function __construct(
+        $message,
+        $code = 0,
+        ?Exception $previous = null,
+        $needle = null,
+        $haystack = null
+    ) {
         // Format the error message
-        $message .= PHP_EOL . PHP_EOL . sprintf(<<<MESSAGE
-================================================================================
-= Needle =======================================================================
-================================================================================
-%s
-
-================================================================================
-= Haystack =====================================================================
-================================================================================
-%s
-
-MESSAGE
+        $message .= PHP_EOL . PHP_EOL . sprintf(
+            <<<MESSAGE
+                ================================================================================
+                = Needle =======================================================================
+                ================================================================================
+                %s
+                
+                ================================================================================
+                = Haystack =====================================================================
+                ================================================================================
+                %s
+                
+                MESSAGE
             ,
-                json_encode($needle, JSON_PRETTY_PRINT),
-                json_encode($haystack, JSON_PRETTY_PRINT)
+            json_encode($needle, JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT),
+            json_encode($haystack, JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT)
         );
 
         parent::__construct($message, $code, $previous);
