@@ -1,12 +1,15 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ServiceContainer;
 
-use Imbo\BehatApiExtension\Context\Initializer\ApiClientAwareInitializer;
-use Imbo\BehatApiExtension\ArrayContainsComparator;
-use Imbo\BehatApiExtension\Context\Initializer\ArrayContainsComparatorAwareInitializer;
 use Behat\Behat\Context\ServiceContainer\ContextExtension;
 use Behat\Testwork\ServiceContainer\Extension as ExtensionInterface;
 use Behat\Testwork\ServiceContainer\ExtensionManager;
+use Imbo\BehatApiExtension\ArrayContainsComparator;
+use Imbo\BehatApiExtension\Context\Initializer\ApiClientAwareInitializer;
+use Imbo\BehatApiExtension\Context\Initializer\ArrayContainsComparatorAwareInitializer;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
@@ -21,39 +24,41 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class BehatApiExtension implements ExtensionInterface {
+class BehatApiExtension implements ExtensionInterface
+{
     /**
      * Service ID for the comparator
      *
      * @var string
      */
-    const COMPARATOR_SERVICE_ID = 'api_extension.comparator';
+    public const COMPARATOR_SERVICE_ID = 'api_extension.comparator';
 
     /**
      * Service ID for the client initializer
      *
      * @var string
      */
-    const APICLIENT_INITIALIZER_SERVICE_ID = 'api_extension.api_client.context_initializer';
+    public const APICLIENT_INITIALIZER_SERVICE_ID = 'api_extension.api_client.context_initializer';
 
     /**
      * Service ID for the initializer
      *
      * @var string
      */
-    const COMPARATOR_INITIALIZER_SERVICE_ID = 'api_extension.comparator.context_initializer';
+    public const COMPARATOR_INITIALIZER_SERVICE_ID = 'api_extension.comparator.context_initializer';
 
     /**
      * Config key for the extension
      *
      * @var string
      */
-    const CONFIG_KEY = 'api_extension';
+    public const CONFIG_KEY = 'api_extension';
 
     /**
      * {@inheritdoc}
      */
-    public function getConfigKey() {
+    public function getConfigKey(): string
+    {
         return self::CONFIG_KEY;
     }
 
@@ -61,38 +66,41 @@ class BehatApiExtension implements ExtensionInterface {
      * {@inheritdoc}
      * @codeCoverageIgnore
      */
-    public function initialize(ExtensionManager $extensionManager) {
+    public function initialize(ExtensionManager $extensionManager): void
+    {
         // Not used
     }
 
     /**
      * {@inheritdoc}
      */
-    public function configure(ArrayNodeDefinition $builder) {
+    public function configure(ArrayNodeDefinition $builder): void
+    {
         $builder
             ->children()
-                ->arrayNode('apiClient')
-                    ->addDefaultsIfNotSet()
-                    ->ignoreExtraKeys(false)
-                    ->children()
-                        ->scalarNode('base_uri')
-                            ->isRequired()
-                            ->cannotBeEmpty()
-                            ->defaultValue('http://localhost:8080')
-                        ->end()
-                    ->end();
+            ->arrayNode('apiClient')
+            ->addDefaultsIfNotSet()
+            ->ignoreExtraKeys(false)
+            ->children()
+            ->scalarNode('base_uri')
+            ->isRequired()
+            ->cannotBeEmpty()
+            ->defaultValue('http://localhost:8080')
+            ->end()
+            ->end();
     }
 
     /**
      * {@inheritdoc}
      * @codeCoverageIgnore
      */
-    public function load(ContainerBuilder $container, array $config) {
+    public function load(ContainerBuilder $container, array $config): void
+    {
         // Client initializer definition
         $clientInitializerDefinition = new Definition(
             ApiClientAwareInitializer::class,
             [
-                $config['apiClient']
+                $config['apiClient'],
             ]
         );
         $clientInitializerDefinition->addTag(ContextExtension::INITIALIZER_TAG);
@@ -104,7 +112,7 @@ class BehatApiExtension implements ExtensionInterface {
         $comparatorInitializerDefinition = new Definition(
             ArrayContainsComparatorAwareInitializer::class,
             [
-                new Reference(self::COMPARATOR_SERVICE_ID)
+                new Reference(self::COMPARATOR_SERVICE_ID),
             ]
         );
         $comparatorInitializerDefinition->addTag(ContextExtension::INITIALIZER_TAG);
@@ -119,7 +127,7 @@ class BehatApiExtension implements ExtensionInterface {
      * {@inheritdoc}
      * @codeCoverageIgnore
      */
-    public function process(ContainerBuilder $container) {
-
+    public function process(ContainerBuilder $container): void
+    {
     }
 }
