@@ -1,15 +1,19 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ServiceContainer;
 
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\Processor;
-use PHPUnit_Framework_TestCase;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\ServiceContainer\BehatApiExtension
  * @testdox Extension
  */
-class BehatApiExtensionTest extends PHPUnit_Framework_TestCase {
+class BehatApiExtensionTest extends TestCase
+{
     /**
      * @var BehatApiExtension
      */
@@ -18,7 +22,8 @@ class BehatApiExtensionTest extends PHPUnit_Framework_TestCase {
     /**
      * Set up the SUT
      */
-    public function setUp() {
+    public function setup(): void
+    {
         $this->extension = new BehatApiExtension();
     }
 
@@ -26,7 +31,8 @@ class BehatApiExtensionTest extends PHPUnit_Framework_TestCase {
      * @covers ::getConfigKey
      * @covers ::configure
      */
-    public function testCanBuildConfiguration() {
+    public function testCanBuildConfiguration()
+    {
         $treeBuilder = new TreeBuilder();
         $rootNode = $treeBuilder->root($this->extension->getConfigKey());
 
@@ -36,17 +42,21 @@ class BehatApiExtensionTest extends PHPUnit_Framework_TestCase {
         // Process the configuration
         $config = (new Processor())->process($rootNode->getNode(true), []);
 
-        $this->assertSame([
-            'apiClient' => [
-                'base_uri' => 'http://localhost:8080',
+        $this->assertSame(
+            [
+                'apiClient' => [
+                    'base_uri' => 'http://localhost:8080',
+                ],
             ],
-        ], $config);
+            $config
+        );
     }
 
     /**
      * @covers ::configure
      */
-    public function testCanOverrideDefaultValuesWhenBuildingConfiguration() {
+    public function testCanOverrideDefaultValuesWhenBuildingConfiguration()
+    {
         $treeBuilder = new TreeBuilder();
         $rootNode = $treeBuilder->root($this->extension->getConfigKey());
 
@@ -54,18 +64,24 @@ class BehatApiExtensionTest extends PHPUnit_Framework_TestCase {
         $this->extension->configure($rootNode);
 
         $baseUri = 'http://localhost:8888';
-        $config = (new Processor())->process($rootNode->getNode(true), [
-            'api_extension' => [
+        $config = (new Processor())->process(
+            $rootNode->getNode(true),
+            [
+                'api_extension' => [
+                    'apiClient' => [
+                        'base_uri' => $baseUri,
+                    ],
+                ],
+            ]
+        );
+
+        $this->assertSame(
+            [
                 'apiClient' => [
                     'base_uri' => $baseUri,
                 ],
             ],
-        ]);
-
-        $this->assertSame([
-            'apiClient' => [
-                'base_uri' => $baseUri,
-            ],
-        ], $config);
+            $config
+        );
     }
 }
